Extract map optimizer ownership
